refactor: update routes and rfc
